dev-tools/requirements.txt
torch
diffusers
Pillow
safetensors
peft
accelerate
